+++ src/index.ts
@@ -1,0 +1,8 @@
+export { CachedNpmInfoClient, NpmPublishClient, UncachedNpmInfoClient } from "./lib/npm-client";
+export { AllPackages } from "./lib/packages";
+export { getLatestTypingVersion } from "./lib/versions";
+
+export { consoleLogger } from "./util/logging";
+export { logUncaughtErrors, nAtATime } from "./util/util";
+
+export { updateLatestTag, updateTypeScriptVersionTags } from "./lib/package-publisher";
